move tokens into styles
