+++ packages/components/src/components/Bar/styles/config.scss
@@ -1,7 +1,7 @@
 // ---------------------------------------------
 // BAR - CONFIGURATION
 // ---------------------------------------------
-@use "../../../tokens";
+@use "../../../styles/tokens";
 
 $spacing: (
   default: tokens.space(),
