+++ packages/components/src/components/Button/styles/config.scss
@@ -3,7 +3,7 @@
 // ---------------------------------------------
 
 @use "sass:map";
-@use "../../../tokens";
+@use "../../../styles/tokens";
 @use "../../../styles/utils";
 @use "./../../Icon/styles/config" as icons;
