+++ packages/components/src/components/Button/styles/mixins.scss
@@ -6,7 +6,7 @@
 @use "sass:list";
 @use "./config";
 @use "../../../styles/utils";
-@use "../../../tokens";
+@use "../../../styles/tokens";
 
 // CONTENTS
 // Base
