+++ packages/components/src/components/Card/styles/config.scss
@@ -1,4 +1,4 @@
-@use "../../../tokens";
+@use "../../../styles/tokens";
 
 $card: (
   default: (
