+++ packages/components/src/components/DropdownMenu/styles/config.scss
@@ -2,7 +2,7 @@
 // DROPDOWN - CONFIGURATION
 // ---------------------------------------------
 
-@use "../../../tokens";
+@use "../../../styles/tokens";
 @use "../../../styles/utils.scss";
 
 $dropdown-menu-max-height: 300px;
