+++ packages/components/src/components/FileUploader/styles/style.scss
@@ -1,4 +1,4 @@
-@use "../../../tokens";
+@use "../../../styles/tokens";
 
 .file-uploader {
   &__preview-item {
