+++ packages/components/src/components/Label/styles/config.scss
@@ -2,7 +2,7 @@
 // FORMS - LABEL - CONFIGURATION
 // ---------------------------------------------
 
-@use "../../../tokens";
+@use "../../../styles/tokens";
 @use "../../../styles/utils";
 
 $label: (
