+++ packages/components/src/components/Modal/styles/config.scss
@@ -1,5 +1,5 @@
 @use "sass:color";
-@use "../../../tokens";
+@use "../../../styles/tokens";
 @use "../../../styles/utils";
 
 $modal-overlay-z-index: 500;
