+++ packages/components/src/components/Modal/styles/mixins.scss
@@ -1,7 +1,7 @@
 @use "sass:map";
 @use "./config";
 @use "../../../styles/utils";
-@use "../../../tokens";
+@use "../../../styles/tokens";
 
 @mixin modal($settings: config.$modal) {
   @each $breakpoint, $props in $settings {
