+++ packages/components/src/components/Modal/styles/style.scss
@@ -1,4 +1,4 @@
-@use "../../../tokens";
+@use "../../../styles/tokens";
 @use "./mixins";
 
 .modal {
