+++ packages/components/src/components/Pagination/styles/style.scss
@@ -1,4 +1,4 @@
-@use "../../../tokens";
+@use "../../../styles/tokens";
 
 .pagination {
   margin-bottom: tokens.space();
