+++ packages/components/src/components/Select/styles/config.scss
@@ -2,7 +2,7 @@
 // FORMS - SELECT - CONFIGURATION
 // ---------------------------------------------
 
-@use "../../../tokens";
+@use "../../../styles/tokens";
 @use "../../../styles/utils";
 
 $select: (
