+++ packages/components/src/components/Table/styles/style.scss
@@ -1,4 +1,4 @@
-@use "../../../tokens";
+@use "../../../styles/tokens";
 @use "../../../styles/utils";
 
 .table {
