+++ packages/components/src/styles/globals.scss
@@ -1,3 +1,3 @@
-@import "../tokens/colors/globals";
-@import "../tokens/space/globals";
-@import "../tokens/font/globals";
+@import "./tokens/colors/globals";
+@import "./tokens/space/globals";
+@import "./tokens/font/globals";
