+++ packages/components/src/styles/tokens/breakpoints/breakpoints.test.ts
@@ -5,7 +5,7 @@
 
 import { renderSync } from "sass";
 import { breakpoint } from ".";
-import scssTestImporter from "../../utils/scssTestImporter";
+import scssTestImporter from "../../../utils/scssTestImporter";
 
 function renderSass(args) {
   return renderSync({
