+++ packages/components/src/styles/tokens/colors/colors.test.ts
@@ -5,7 +5,7 @@
 
 import { renderSync } from "sass";
 import { color, colorHex } from ".";
-import scssTestImporter from "../../utils/scssTestImporter";
+import scssTestImporter from "../../../utils/scssTestImporter";
 
 function renderSass(args) {
   return renderSync({
