+++ packages/components/src/styles/tokens/colors/globals.scss
