+++ packages/components/src/styles/tokens/font/index.ts
@@ -1,4 +1,4 @@
-import scssVarsToMap from "../../utils/scssVarsToMap";
+import scssVarsToMap from "../../../utils/scssVarsToMap";
 import fontsSCSSVars from "./export.module.scss";
 
 export const fonts = scssVarsToMap(fontsSCSSVars);
