+++ packages/components/src/styles/tokens/shadows/index.ts
@@ -1,4 +1,4 @@
-import scssVarsToMap from "../../utils/scssVarsToMap";
+import scssVarsToMap from "../../../utils/scssVarsToMap";
 import shadowsSCSSVars from "./export.module.scss";
 
 export const shadows = scssVarsToMap(shadowsSCSSVars);
